Massive refactoring
A lot of decoupled code.
