+++ src/utilities.ts
@@ -1,3 +1,5 @@
+import { User } from './db/user'
+
 // All types of buttons available on the page
 export enum ButtonType {
     Search,
@@ -12,13 +14,6 @@ export const getResultField = () => document.getElementById("queryResult")
 export const getInputField = () => document.getElementById("inputField") as HTMLInputElement
 
 export const makeP = text => "<p>" + text + "</p>"
-export const makeTBodyEl = (dataN, dataId, dataDOB, dataPDV) => {
-    "<td>" + dataN + "</td>"
-    "<td>" + dataId + "</td>"
-    "<td>" + dataDOB + "</td>"
-    "<td>" + dataPDV + "</td>"
-
-}
 
 export const getButtonByType = (type: ButtonType): HTMLAnchorElement => {
     let buttonID = "";
@@ -53,3 +48,32 @@ export const clearPage = () => {
 
 export const setSubmitButtonVisibility = (isVisible: boolean) =>
     getButtonByType(ButtonType.Submit).style.display = isVisible ? "inline-block" : "none"
+
+export type OnClickFn = (this: GlobalEventHandlers, ev: MouseEvent) => any
+
+
+const bindButtonEvent = (btnType: ButtonType, fn: OnClickFn) =>
+    getButtonByType(btnType).onclick = fn
+
+export const bindButtons = (bindings: Array<[ButtonType, OnClickFn]>) => {
+    for (let [btn, fn] of bindings) {
+        bindButtonEvent(btn, fn)
+    }
+}
+
+
+export const fillUsersTable = (users: Array<User>) => {
+    const table = document.getElementById("tInfoB")
+    table.innerHTML = ""
+    for (const user of users) {
+        const userName = user.name
+        const dateOfBAsStr = (user.yearOfBirth).toString()
+        const pavelDVisitsAsStr = (user.pavelDurovVisits).toString()
+        const idAsStr = users.indexOf(user)
+        table.innerHTML += table.appendChild(document.createElement("tr")).innerHTML =
+            '<td class="col s3">' + userName + "</td>" +
+            '<td class="col s3">' + idAsStr + "</td>" +
+            '<td class="col s3">' + dateOfBAsStr + "</td>" +
+            '<td class="col s3">' + pavelDVisitsAsStr + "</td>"
+    }
+}
